# [11_17] Use local font_rev_table to speedup tt_font_exists