--- a/src/Graphics/Fonts/font_database.cpp
+++ b/src/Graphics/Fonts/font_database.cpp
@@ -80,7 +80,9 @@ bool                  new_fonts          = false;
 static bool           fonts_loaded       = false;
 static bool           fonts_global_loaded= false;
 hashmap<tree, tree>   font_table (UNINIT);
+hashmap<string, tree> font_rev_table (UNINIT);
 hashmap<tree, tree>   font_global_table (UNINIT);
+hashmap<string, tree> font_global_rev_table (UNINIT);
 hashmap<tree, tree>   font_features (UNINIT);
 hashmap<tree, tree>   font_variants (UNINIT);
 hashmap<tree, tree>   font_characteristics (UNINIT);
@@ -103,7 +105,8 @@ tuple_insert (tree& t, tree x) {
 }
 
 void
-font_database_load_database (url u, hashmap<tree, tree>& ftab= font_table) {
+font_database_load_database (url u, hashmap<tree, tree>& ftab= font_table,
+                             hashmap<string, tree>& frev= font_rev_table) {
   if (!exists (u)) return;
   string s;
   if (!load_string (u, s, false)) {
@@ -111,8 +114,12 @@ font_database_load_database (url u, hashmap<tree, tree>& ftab= font_table) {
     for (int i= 0; i < N (t); i++)
       if (is_func (t[i], TUPLE, 2)) {
         // if (&ftab == &font_table)
-        //   cout << t[i][0] << " ~> " << t[i][1] << "\n";
-        ftab (t[i][0])= t[i][1];
+        tree family_style  = t[i][0];
+        tree files         = t[i][1];
+        ftab (family_style)= files;
+        for (int j= 0; j < N (files); j++) {
+          frev (files[j][0]->label)= family_style;
+        }
       }
   }
 }
@@ -244,7 +251,8 @@ void
 font_database_global_load () {
   if (fonts_global_loaded) return;
   cout << "TeXmacs] warning, missing font, loading global substitution list\n";
-  font_database_load_database (GLOBAL_DATABASE, font_global_table);
+  font_database_load_database (GLOBAL_DATABASE, font_global_table,
+                               font_global_rev_table);
   font_database_load_features (GLOBAL_FEATURES);
   font_database_load_characteristics (GLOBAL_CHARACTERISTICS);
   font_database_load_substitutions (GLOBAL_SUBSTITUTIONS);
@@ -710,4 +718,10 @@ font_database_substitutions (string family) {
   font_database_load ();
   if (font_substitutions->contains (family)) return font_substitutions[family];
   else return tree (TUPLE);
+}
+
+bool
+font_database_rev_contains (string font_file_name) {
+  font_database_load ();
+  return font_rev_table->contains (font_file_name);
 }