Before
```
TeXmacs] std-bench, Task 'tt_locate Songti.pfb' took 1057 ms
TeXmacs] std-bench, Task 'tt_locate Songti.ttf' took 23 ms
TeXmacs] std-bench, Task 'tt_locate Songti.ttc' took 19 ms
TeXmacs] std-bench, Task 'tt_font_exists Songti' took 1099 ms
```
Now
```
TeXmacs] std-bench, Task 'tt_locate Songti.ttc' took 16 ms
TeXmacs] std-bench, Task 'tt_font_exists Songti' took 270 ms
```
or
```
TeXmacs] std-bench, Task 'tt_locate Songti.ttc' took 17 ms
TeXmacs] std-bench, Task 'tt_font_exists Songti' took 36 ms
```